+++ wgpu-hal/src/metal/adapter.rs
@@ -47,22 +47,11 @@ fn device_class_responds_to(device: &ProtocolObject<dyn MTLDevice>, sel: Sel) ->
 /// [new command buffer]: https://developer.apple.com/documentation/metal/mtlcommandqueue/makecommandbuffer()?language=objc
 pub(super) const MAX_COMMAND_BUFFERS: usize = 4096;
 
-// Metal has a single buffer limit that we must split across 3 WebGPU limits:
-// The Metal limit is: 31 "Maximum number of entries in the buffer argument table, per graphics or kernel function".
-// We must split it across:
-//  - maxStorageBuffersPerShaderStage; must be at least 8
-//  - maxUniformBuffersPerShaderStage; must be at least 12
-//  - maxVertexBuffers; must be at least 8
-// We require 2 additional internal buffers:
-//  - one for immediate data
-//  - one for sizes of other buffers
-// We use the last buffer for an acceleration structure.
-const MAX_STORAGE_BUFFERS_PER_SHADER_STAGE: u32 = 8;
-const MAX_UNIFORM_BUFFERS_PER_SHADER_STAGE: u32 = 12;
-const MAX_VERTEX_BUFFERS: u32 = 8;
-const MAX_ACCELERATION_STRUCTURES_PER_SHADER_STAGE: u32 = 1;
-// Use the end of the range for vertex buffers.
-pub const VERTEX_BUFFER_SLOT_START: u32 = 31 - 8;
+/// "Maximum number of entries in the buffer argument table, per graphics or kernel function"
+///
+/// Vertex buffers are placed at the end of the table (highest indices),
+/// counting down from MAX_BUFFERS - 1.
+pub const MAX_BUFFERS: u32 = 31;
 
 impl super::Adapter {
     pub(super) fn new(shared: Arc<super::AdapterShared>) -> Self {
@@ -1341,7 +1330,7 @@ impl super::CapabilitiesQuery {
         features
     }
 
-    pub fn capabilities(&self) -> crate::Capabilities {
+    pub fn capabilities(&self, instance_flags: wgt::InstanceFlags) -> crate::Capabilities {
         let mut downlevel = wgt::DownlevelCapabilities::default();
         downlevel.flags.set(
             wgt::DownlevelFlags::FRAGMENT_WRITABLE_STORAGE,
@@ -1378,6 +1367,34 @@ impl super::CapabilitiesQuery {
             self.format_bc || (self.format_eac_etc && self.format_astc),
         );
 
+        let max_storage_buffers_per_shader_stage;
+        let max_uniform_buffers_per_shader_stage;
+        let max_vertex_buffers;
+        let max_acceleration_structures_per_shader_stage;
+        let max_buffers_and_acceleration_structures_per_shader_stage;
+
+        // Metal has a single buffer limit that we must split across 3 WebGPU limits:
+        //  - maxStorageBuffersPerShaderStage; must be at least 8
+        //  - maxUniformBuffersPerShaderStage; must be at least 12
+        //  - maxVertexBuffers; must be at least 8
+        // We also have to reserve 2 additional internal buffers:
+        //  - one for immediate data
+        //  - one for sizes of other buffers
+        if instance_flags.contains(wgt::InstanceFlags::STRICT_WEBGPU_COMPLIANCE) {
+            max_storage_buffers_per_shader_stage = 9;
+            max_uniform_buffers_per_shader_stage = 12;
+            max_vertex_buffers = 8;
+            max_acceleration_structures_per_shader_stage = 0;
+            max_buffers_and_acceleration_structures_per_shader_stage = u32::MAX;
+        } else {
+            const MAX_USABLE_BUFFERS: u32 = MAX_BUFFERS - 2;
+            max_storage_buffers_per_shader_stage = MAX_USABLE_BUFFERS;
+            max_uniform_buffers_per_shader_stage = MAX_USABLE_BUFFERS;
+            max_vertex_buffers = MAX_USABLE_BUFFERS;
+            max_acceleration_structures_per_shader_stage = MAX_USABLE_BUFFERS;
+            max_buffers_and_acceleration_structures_per_shader_stage = MAX_USABLE_BUFFERS;
+        }
+
         let limits = crate::auxil::adjust_raw_limits(wgt::Limits {
             //
             // WebGPU LIMITS:
@@ -1394,16 +1411,16 @@ impl super::CapabilitiesQuery {
             // No limit.
             max_bindings_per_bind_group: u32::MAX,
             // No limit, use maxUniformBuffersPerShaderStage.
-            max_dynamic_uniform_buffers_per_pipeline_layout: MAX_UNIFORM_BUFFERS_PER_SHADER_STAGE,
+            max_dynamic_uniform_buffers_per_pipeline_layout: max_uniform_buffers_per_shader_stage,
             // No limit, use maxStorageBuffersPerShaderStage.
-            max_dynamic_storage_buffers_per_pipeline_layout: MAX_STORAGE_BUFFERS_PER_SHADER_STAGE,
+            max_dynamic_storage_buffers_per_pipeline_layout: max_storage_buffers_per_shader_stage,
             // "Maximum number of entries in the sampler state argument table, per graphics or kernel function"
             max_samplers_per_shader_stage: 16,
             max_sampled_textures_per_shader_stage: self.max_textures_per_stage.0,
             max_storage_textures_per_shader_stage: self.max_textures_per_stage.1,
-            max_storage_buffers_per_shader_stage: MAX_STORAGE_BUFFERS_PER_SHADER_STAGE,
-            max_uniform_buffers_per_shader_stage: MAX_UNIFORM_BUFFERS_PER_SHADER_STAGE,
-            max_vertex_buffers: MAX_VERTEX_BUFFERS,
+            max_storage_buffers_per_shader_stage,
+            max_uniform_buffers_per_shader_stage,
+            max_vertex_buffers,
             max_buffer_size: self.max_buffer_size,
             // No limit, use maxBufferSize.
             max_uniform_buffer_binding_size: self.max_buffer_size,
@@ -1445,8 +1462,8 @@ impl super::CapabilitiesQuery {
             // From 2.17.7 in https://developer.apple.com/metal/Metal-Shading-Language-Specification.pdf
             // > [Acceleration structures] are opaque objects that can be bound directly using
             // buffer binding points or via argument buffers
-            max_acceleration_structures_per_shader_stage:
-                MAX_ACCELERATION_STRUCTURES_PER_SHADER_STAGE,
+            max_acceleration_structures_per_shader_stage,
+            max_buffers_and_acceleration_structures_per_shader_stage,
 
             max_multiview_view_count: if self.supported_vertex_amplification_factor > 1 {
                 self.supported_vertex_amplification_factor