+++ wgpu-types/src/limits.rs
@@ -87,6 +87,10 @@ macro_rules! with_limits {
         $macro_name!(max_blas_geometry_count, Ordering::Less);
         $macro_name!(max_tlas_instance_count, Ordering::Less);
         $macro_name!(max_acceleration_structures_per_shader_stage, Ordering::Less);
+        $macro_name!(
+            max_buffers_and_acceleration_structures_per_shader_stage,
+            Ordering::Less
+        );
 
         $macro_name!(max_multiview_view_count, Ordering::Less);
 
@@ -316,6 +320,9 @@ pub struct Limits {
     /// Requesting more than 0 during device creation only makes sense if [`Features::EXPERIMENTAL_RAY_QUERY`]
     /// is enabled.
     pub max_acceleration_structures_per_shader_stage: u32,
+    /// The combined number of buffers (storage and uniform), vertex buffers, and acceleration
+    /// structures that can be bound in a single shader stage.
+    pub max_buffers_and_acceleration_structures_per_shader_stage: u32,
 
     /// The maximum number of views that can be used in multiview rendering
     pub max_multiview_view_count: u32,
@@ -401,6 +408,7 @@ impl Limits {
     ///     max_blas_geometry_count: 0,
     ///     max_tlas_instance_count: 0,
     ///     max_acceleration_structures_per_shader_stage: 0,
+    ///     max_buffers_and_acceleration_structures_per_shader_stage: 28, // sum of storage buffers, uniform buffers and vertex buffers limits
     ///     max_multiview_view_count: 0,
     ///     max_ray_dispatch_count: 0,
     ///     max_ray_recursion_depth: 0,
@@ -467,6 +475,7 @@ impl Limits {
             max_blas_geometry_count: 0,
             max_tlas_instance_count: 0,
             max_acceleration_structures_per_shader_stage: 0,
+            max_buffers_and_acceleration_structures_per_shader_stage: 28,
 
             max_multiview_view_count: 0,
 
@@ -536,6 +545,7 @@ impl Limits {
     ///     max_blas_geometry_count: 0,
     ///     max_tlas_instance_count: 0,
     ///     max_acceleration_structures_per_shader_stage: 0,
+    ///     max_buffers_and_acceleration_structures_per_shader_stage: 24, // * sum of storage buffers, uniform buffers and vertex buffers limits
     ///
     ///     max_multiview_view_count: 0,
     ///
@@ -555,6 +565,7 @@ impl Limits {
             max_color_attachments: 4,
             // see: https://developer.apple.com/metal/Metal-Feature-Set-Tables.pdf#page=7
             max_compute_workgroup_storage_size: 16352,
+            max_buffers_and_acceleration_structures_per_shader_stage: 24,
             ..Self::defaults()
         }
     }
@@ -621,6 +632,7 @@ impl Limits {
     ///     max_blas_geometry_count: 0,
     ///     max_tlas_instance_count: 0,
     ///     max_acceleration_structures_per_shader_stage: 0,
+    ///     max_buffers_and_acceleration_structures_per_shader_stage: 19, // * sum of storage buffers, uniform buffers and vertex buffers limits
     ///
     ///     max_multiview_view_count: 0,
     ///
@@ -646,6 +658,8 @@ impl Limits {
 
             // Value supported by Intel Celeron B830 on Windows (OpenGL 3.1)
             max_inter_stage_shader_variables: 15,
+
+            max_buffers_and_acceleration_structures_per_shader_stage: 19,
 
             // Most of the values should be the same as the downlevel defaults
             ..Self::downlevel_defaults()
@@ -722,6 +736,7 @@ impl Limits {
             max_blas_geometry_count: ALLOC_MAX_U32,
             max_tlas_instance_count: ALLOC_MAX_U32,
             max_acceleration_structures_per_shader_stage: ALLOC_MAX_U32,
+            max_buffers_and_acceleration_structures_per_shader_stage: ALLOC_MAX_U32,
 
             max_multiview_view_count: ALLOC_MAX_U32,
             max_ray_dispatch_count: ALLOC_MAX_U32,
@@ -765,6 +780,7 @@ impl Limits {
             max_blas_primitive_count: 1 << 28,      // 2^28: Metal's minimum
             // On metal acceleration structures are limited because they share buffer slots
             max_acceleration_structures_per_shader_stage: 1,
+            max_buffers_and_acceleration_structures_per_shader_stage: 29,
             ..self
         }
     }
@@ -779,6 +795,8 @@ impl Limits {
             max_blas_primitive_count: other.max_blas_primitive_count,
             max_acceleration_structures_per_shader_stage: other
                 .max_acceleration_structures_per_shader_stage,
+            max_buffers_and_acceleration_structures_per_shader_stage: other
+                .max_buffers_and_acceleration_structures_per_shader_stage,
             ..self
         }
     }